Supabase account
The dotenv.env file has the url, key, password etc for a demo project. Modified how the supabase is loaded in the main cause it wasnt working on the branch i started from. Also created a simple user.py. The others remain the same. Note that im using my local path on main.py for loading the dotenv. the dotenv file should be in the same folder as the main to execute correctly.